pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 110
skip-string-normalization = true
target-version = ['py36', 'py37', 'py38']

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version = "0.2.0"
version_files = [
  "pyproject.toml:version",
  "src/mps_client/__init__.py",
]

[tool.poetry]
authors = ["Michael Statt <michael.statt@modelyst.io>"]
description = "MPS Plotting and CLI Tool"
license = "Apache-2.0"
name = "mps-client"
version = "0.2.0"

[tool.poetry.dependencies]
aiofiles = "^22.1.0"
beautifulsoup4 = "^4.11.1"
httpx = "^0.23.0"
matplotlib = "^3.6.1"
modelyst-dbgen = {version = "^1.0.0a7", allow-prereleases = true}
pandas = "^1.5.1"
python = ">=3.8.1,<4.0"
requests = "^2.28.1"
typer = "^0.6.1"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.7.6"
black = "^22.10.0"
commitizen = "^2.35.0"
flake8 = "^5.0.4"
ipykernel = "^6.17.1"
isort = "^5.10.1"
mypy = "^0.982"
pre-commit = "^2.20.0"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-sugar = "^0.9.5"
types-requests = "^2.28.11.2"

[tool.poetry.scripts]
mps-client = "mps_client.__main__:main"

[tool.pytest.ini_options]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = [
  "tests",
]
